requirements.txt
bcrypt==4.2.0
blinker==1.8.2
click==8.1.7
exceptiongroup==1.2.2
Flask==3.0.3
importlib_metadata==8.2.0
iniconfig==2.0.0
itsdangerous==2.2.0
Jinja2==3.1.4
MarkupSafe==2.1.5
packaging==24.1
pluggy==1.5.0
PyJWT==2.9.0
pytest==8.3.2
python-dotenv==1.0.1
tomli==2.0.1
Werkzeug==3.0.3
zipp==3.20.0
